requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.0
colorama==0.4.6
Django==5.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-taggit==5.0.1
djangorestframework==3.16.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
Markdown==3.8
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2025.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.4.0
uvicorn==0.34.2
whitenoise==6.9.0
